+++ src/Payum/Core/PayumBuilder.php
@@ -109,14 +109,11 @@ class PayumBuilder
     protected $mainRegistry;
 
     /**
+     * @deprecated will be removed in 2.0
+     *
      * @var HttpClientInterface
      */
     protected $httpClient;
-
-    /**
-     * @var string
-     */
-    protected $modelIdProperty;
 
     /**
      * @return static
@@ -363,6 +360,8 @@ class PayumBuilder
     /**
      * @param HttpClientInterface $httpClient
      *
+     * @deprecated this method will be removed in 2.0 Use self::addCoreGatewayFactoryConfig to overwrite http client.
+     *
      * @return static
      */
     public function setHttpClient(HttpClientInterface $httpClient = null)
@@ -393,14 +392,9 @@ class PayumBuilder
 
         $httpRequestVerifier = $this->buildHttpRequestVerifier($this->tokenStorage);
 
-        if (false == $httpClient = $this->httpClient) {
-            $httpClient = HttpClientFactory::create();
-        }
-
         $coreGatewayFactory = $this->buildCoreGatewayFactory(array_replace_recursive([
             'payum.extension.token_factory' => new GenericTokenFactoryExtension($genericTokenFactory),
             'payum.security.token_storage' => $tokenStorage,
-            'payum.http_client' => $httpClient,
         ], $this->coreGatewayFactoryConfig));
 
         $gatewayFactories = array_replace(