+++ src/Payum/Paypal/ExpressCheckout/Nvp/Api.php
@@ -1,8 +1,7 @@
 <?php
 namespace Payum\Paypal\ExpressCheckout\Nvp;
 
-use GuzzleHttp\Psr7\Request;
-use Payum\Core\Bridge\Guzzle\HttpClientFactory;
+use Http\Message\MessageFactory;
 use Payum\Core\Bridge\Spl\ArrayObject;
 use Payum\Core\Exception\Http\HttpException;
 use Payum\Core\Exception\InvalidArgumentException;
@@ -289,7 +288,15 @@ class Api
 
     const VERSION = '65.1';
 
+    /**
+     * @var HttpClientInterface
+     */
     protected $client;
+
+    /**
+     * @var MessageFactory
+     */
+    protected $messageFactory;
 
     protected $options = array(
         'username' => null,
@@ -303,10 +310,11 @@ class Api
     );
 
     /**
-     * @param array                $options
+     * @param array                    $options
      * @param HttpClientInterface|null $client
+     * @param MessageFactory|null      $messageFactory
      */
-    public function __construct(array $options, HttpClientInterface $client = null)
+    public function __construct(array $options, HttpClientInterface $client, MessageFactory $messageFactory)
     {
         $options = ArrayObject::ensureArrayObject($options);
         $options->defaults($this->options);
@@ -321,7 +329,8 @@ class Api
         }
 
         $this->options = $options;
-        $this->client = $client ?: HttpClientFactory::create();
+        $this->client = $client;
+        $this->messageFactory = $messageFactory;
     }
 
     /**
@@ -532,7 +541,7 @@ class Api
             'Content-Type' => 'application/x-www-form-urlencoded',
         );
 
-        $request = new Request('POST', $this->getApiEndpoint(), $headers, http_build_query($fields));
+        $request = $this->messageFactory->createRequest('POST', $this->getApiEndpoint(), $headers, http_build_query($fields));
 
         $response = $this->client->send($request);