+++ src/Payum/Paypal/ProCheckout/Nvp/PaypalProCheckoutGatewayFactory.php
@@ -49,7 +49,7 @@ class PaypalProCheckoutGatewayFactory extends GatewayFactory
                     'sandbox' => $config['sandbox'],
                 );
 
-                return new Api($paypalConfig, $config['payum.http_client']);
+                return new Api($paypalConfig, $config['payum.http_client'], $config['httplug.message_factory']);
             };
         }
     }