+++ src/Payum/Skeleton/Api.php
@@ -1,8 +1,7 @@
 <?php
 namespace Payum\Skeleton;
 
-use GuzzleHttp\Psr7\Request;
-use Payum\Core\Bridge\Guzzle\HttpClientFactory;
+use Http\Message\MessageFactory;
 use Payum\Core\Exception\Http\HttpException;
 use Payum\Core\HttpClientInterface;
 
@@ -14,6 +13,11 @@ class Api
     protected $client;
 
     /**
+     * @var MessageFactory
+     */
+    protected $messageFactory;
+
+    /**
      * @var array
      */
     protected $options = [];
@@ -21,13 +25,15 @@ class Api
     /**
      * @param array               $options
      * @param HttpClientInterface $client
+     * @param MessageFactory      $messageFactory
      *
      * @throws \Payum\Core\Exception\InvalidArgumentException if an option is invalid
      */
-    public function __construct(array $options, HttpClientInterface $client = null)
+    public function __construct(array $options, HttpClientInterface $client, MessageFactory $messageFactory)
     {
         $this->options = $options;
-        $this->client = $client ?: HttpClientFactory::create();
+        $this->client = $client;
+        $this->messageFactory = $messageFactory;
     }
 
     /**
@@ -39,7 +45,7 @@ class Api
     {
         $headers = [];
 
-        $request = new Request($method, $this->getApiEndpoint(), $headers, http_build_query($fields));
+        $request = $this->messageFactory->createRequest($method, $this->getApiEndpoint(), $headers, http_build_query($fields));
 
         $response = $this->client->send($request);