+++ src/Payum/Skeleton/SkeletonGatewayFactory.php
@@ -40,7 +40,7 @@ class SkeletonGatewayFactory extends GatewayFactory
             $config['payum.api'] = function (ArrayObject $config) {
                 $config->validateNotEmpty($config['payum.required_options']);
 
-                return new Api((array) $config, $config['payum.http_client']);
+                return new Api((array) $config, $config['payum.http_client'], $config['httplug.message_factory']);
             };
         }
     }