+++ src/Payum/Stripe/Action/Api/CreateCustomerAction.php
@@ -4,13 +4,10 @@ namespace Payum\Stripe\Action\Api;
 use Payum\Core\Action\GatewayAwareAction;
 use Payum\Core\ApiAwareInterface;
 use Payum\Core\Bridge\Spl\ArrayObject;
-use Payum\Core\Exception\LogicException;
 use Payum\Core\Exception\RequestNotSupportedException;
 use Payum\Core\Exception\UnsupportedApiException;
 use Payum\Stripe\Keys;
-use Payum\Stripe\Request\Api\CreateCharge;
 use Payum\Stripe\Request\Api\CreateCustomer;
-use Payum\Stripe\Request\Api\ObtainToken;
 use Stripe\Customer;
 use Stripe\Error;
 use Stripe\Stripe;
@@ -39,7 +36,7 @@ class CreateCustomerAction extends GatewayAwareAction implements ApiAwareInterfa
      */
     public function execute($request)
     {
-        /** @var $request CreateCharge */
+        /** @var $request CreateCustomer */
         RequestNotSupportedException::assertSupports($this, $request);
 
         $model = ArrayObject::ensureArrayObject($request->getModel());
@@ -47,10 +44,10 @@ class CreateCustomerAction extends GatewayAwareAction implements ApiAwareInterfa
         try {
             Stripe::setApiKey($this->keys->getSecretKey());
 
-            $charge = Customer::create($model->toUnsafeArrayWithoutLocal());
+            $customer = Customer::create($model->toUnsafeArrayWithoutLocal());
 
-            $model->replace($charge->__toArray(true));
-        } catch (Error\Card $e) {
+            $model->replace($customer->__toArray(true));
+        } catch (Error\Base $e) {
             $model->replace($e->getJsonBody());
         }
     }