+++ src/Payum/Stripe/Action/Api/CreatePlanAction.php
@@ -1,19 +1,18 @@
 <?php
 namespace Payum\Stripe\Action\Api;
 
-use Payum\Core\Action\ActionInterface;
+use Payum\Core\Action\GatewayAwareAction;
 use Payum\Core\ApiAwareInterface;
 use Payum\Core\Bridge\Spl\ArrayObject;
-use Payum\Core\Exception\LogicException;
 use Payum\Core\Exception\RequestNotSupportedException;
 use Payum\Core\Exception\UnsupportedApiException;
 use Payum\Stripe\Keys;
-use Payum\Stripe\Request\Api\CreateCharge;
-use Stripe\Charge;
+use Payum\Stripe\Request\Api\CreatePlan;
 use Stripe\Error;
+use Stripe\Plan;
 use Stripe\Stripe;
 
-class CreateChargeAction implements ActionInterface, ApiAwareInterface
+class CreatePlanAction extends GatewayAwareAction implements ApiAwareInterface
 {
     /**
      * @var Keys
@@ -37,25 +36,17 @@ class CreateChargeAction implements ActionInterface, ApiAwareInterface
      */
     public function execute($request)
     {
-        /** @var $request CreateCharge */
+        /** @var $request CreatePlan */
         RequestNotSupportedException::assertSupports($this, $request);
 
         $model = ArrayObject::ensureArrayObject($request->getModel());
 
-        if (false == ($model['card'] || $model['customer'])) {
-            throw new LogicException('The either card token or customer id has to be set.');
-        }
-
-        if (is_array($model['card'])) {
-            throw new LogicException('The token has already been used.');
-        }
-
         try {
             Stripe::setApiKey($this->keys->getSecretKey());
 
-            $charge = Charge::create($model->toUnsafeArrayWithoutLocal());
+            $plan = Plan::create($model->toUnsafeArrayWithoutLocal());
 
-            $model->replace($charge->__toArray(true));
+            $model->replace($plan->__toArray(true));
         } catch (Error\Base $e) {
             $model->replace($e->getJsonBody());
         }
@@ -67,7 +58,7 @@ class CreateChargeAction implements ActionInterface, ApiAwareInterface
     public function supports($request)
     {
         return
-            $request instanceof CreateCharge &&
+            $request instanceof CreatePlan &&
             $request->getModel() instanceof \ArrayAccess
         ;
     }