+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 1.2.8 (2016-03-25)
+
+* [stripe] Allow use of explicitly passed credit card.
+* [bridge][symfony] Add Symfony's template engine aware `RenderTemplateAction` action.
+
+## 1.2.7 (2016-03-24)
+
+* [stripe] Subscription billing.
+
 ## 1.2.6 (2016-03-18)
 
 * [bridge][symfony] add obtain credit card action builder.